Move FlatReasonsForRejectionPresenter into support interface
